# src/main/resources/templates/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Connect new sites to the OrganiCity platform, as well as perform the management of its related attributes",
		"version": "1.0.0",
		"title": "OrganiCity Site Management API",
		"contact": {"email": "etienne.gandrille@cea.fr"}
	},
	"host": "{{host}}:{{exposedport}}",
	"basePath": "/v1",
...
	"schemes": ["{{protocol}}"],
	"paths": {
		"/sites": {
			"get": {
				"tags": ["site"],
				"summary": "Finds all the sites",
				"operationId": "getSites",
				"produces": ["application/json"],
				"parameters": [{
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token (no role required)",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "An array of OC site summaries",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/sitesummary"
							}
						}
					},
					"401": {
						"description": "Authorization error"
					}
				}
			}
		},
		"/sites/{siteName}": {
			"get": {
				"tags": ["site"],
				"summary": "Finds site by name",
				"operationId": "getSiteByName",
				"produces": ["application/json"],
				"parameters": [{
					"name": "siteName",
					"in": "path",
					"description": "Name of site to return",
					"required": true,
					"type": "string"
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:site-{siteName}-user",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Returns a single site",
						"schema": {
							"$ref": "#/definitions/sitedesc"
						}
					},
					"400": {
						"description": "Invalid name supplied"
					},
					"401": {
						"description": "Authorization error"
					},
					"404": {
						"description": "Site not found"
					}
				}
			},
			"put": {
				"tags": ["site"],
				"summary": "Updates an existing site",
				"operationId": "updateSite",
				"consumes": [
					"application/json"
				],
				"produces": ["application/json"],
				"parameters": [{
					"name": "siteName",
					"in": "path",
					"description": "Name of site that needs to be updated",
					"required": true,
					"type": "string"
				}, {
					"name": "site",
					"in": "body",
					"description": "Site object that needs to be updated",
					"required": true,
					"schema": {
						"$ref": "#/definitions/siteupdate"
					}
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:site-{siteName}-admin",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Site updated",
						"schema": {
							"$ref": "#/definitions/sitedesc"
						}
					},
					"400": {
						"description": "Malformed or invalid json content (i.e. missing fields), or invalid name supplied"
					},
					"401": {
						"description": "Authorization error"
					},
					"404": {
						"description": "Site not found"
					}
				}
			}
		},
		"/sites/{siteName}/quota/increment": {
			"get": {
				"tags": ["site"],
				"summary": "Increment the remaining quota of a site",
				"operationId": "incrementRemainingQuota",
				"produces": ["application/json"],
				"parameters": [{
					"name": "siteName",
					"in": "path",
					"description": "Site name",
					"required": true,
					"type": "string"
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:site-{siteName}-admin",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Quota updated",
						"schema": {
							"$ref": "#/definitions/sitedesc"
						}
					},
					"400": {
						"description": "Invalid site name supplied"
					},
					"401": {
						"description": "Authorization error"
					},
					"403": {
						"description": "Remaining quota is already set to its maximum"
					},
					"404": {
						"description": "Site not found"
					}
				}
			}
		},	
		"/sites/{siteName}/quota/decrement": {
			"get": {
				"tags": ["site"],
				"summary": "Decrement the remaining quota of a site",
				"operationId": "decrementRemainingQuota",
				"produces": ["application/json"],
				"parameters": [{
					"name": "siteName",
					"in": "path",
					"description": "Site name",
					"required": true,
					"type": "string"
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:site-{siteName}-admin",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Quota updated",
						"schema": {
							"$ref": "#/definitions/sitedesc"
						}
					},
					"400": {
						"description": "Invalid site name supplied"
					},
					"401": {
						"description": "Authorization error"
					},
					"403": {
						"description": "Remaining quota is already zero"
					},
					"404": {
						"description": "Site not found"
					}
				}
			}
		},
...
		"/quota": {
			"get": {
				"tags": ["quota"],
				"summary": "Gets the default quota for new experiments",
				"operationId": "getQuota",
				"produces": ["application/json"],
				"parameters": [{
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token (no role required)",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Success",
						"schema": {
							"$ref": "#/definitions/quotadesc"
						}
					},
					"401": {
						"description": "Authorization error"
					}
				}
			}
		},
		"/dictionary/unregisteredassettype": {
			"get": {
				"tags": ["dictionary"],
				"summary": "Finds all the unregistered asset types",
				"operationId": "getUnregisteredAssetTypes",
				"produces": ["application/json"],
				"parameters": [{
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:dictionary-user",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "An array of OC unregistered asset types",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/unregisteredassettypedesc"
							}
						}
					},
					"401": {
						"description": "Authorization error"
					}
				}
			},
			"post": {
				"tags": ["dictionary"],
				"summary": "Add a new unregistered asset type",
				"operationId": "addUnregisteredAssetType",
				"consumes": [
					"application/json"
				],
				"produces": ["application/json"],
				"parameters": [{
					"in": "body",
					"name": "unregisteredAssetType",
					"description": "Unregistered asset type object that needs to be added",
					"required": true,
					"schema": {
						"$ref": "#/definitions/unregisteredassettypeupdate"
					}
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:site-{siteName}-admin",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Unregistered asset type added (or already added before this call)",
						"schema": {
							"$ref": "#/definitions/unregisteredassettypedesc"
						}
					},
					"400": {
						"description": "Malformed json"
					},
					"401": {
						"description": "Authorization error"
					},
					"405": {
						"description": "Invalid json content (i.e. missing fields)"
					},
					"412": {
						"description": "Already existing data type in the data type data base. It does NOT make sense to register it as an unknown data type"
					}
				}
			}
		},
...
		"/dictionary/units": {
			"get": {
				"tags": ["dictionary"],
				"summary": "Finds all the units",
				"operationId": "getUnits",
				"produces": ["application/json"],
				"parameters": [{
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:dictionary-user",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "An array of OC units",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/unitdesc"
							}
						}
					},
					"401": {
						"description": "Authorization error"
					}
				}
			}
		},
		"/dictionary/units/{name}": {
			"get": {
				"tags": ["dictionary"],
				"summary": "Find unit by name",
				"operationId": "getUnitByName",
				"produces": ["application/json"],
				"parameters": [{
					"name": "name",
					"in": "path",
					"description": "unit name",
					"required": true,
					"type": "string"
				}, {
					"name": "Authorization",
					"in": "header",
					"description": "Authorization token with role {{clientId}}:dictionary-user",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/unitdesc"
						}
					},
					"401": {
						"description": "Authorization error"
					},
					"404": {
						"description": "Unit not found"
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"sitesummary": {
			"type": "object",
			"required": [
				"urn",
				"name"
			],
			"properties": {
				"urn": {
					"type": "string",
					"description": "unique urn representing a specific OC site"
				},
				"name": {
					"type": "string",
					"description": "unique identifier representing a specific OC site"
				}
			}
		},
		"siteupdate": {
			"type": "object",
			"required": [
				"email",
				"related"
			],
			"properties": {
				"email": {
					"type": "string",
					"format": "email",
					"description": "administrator's email address"
				},

				"related": {
					"type": "string",
					"description": "extension point for storing structured or unstructured information"
				}
			}
		},
		"sitedesc": {
			"type": "object",
			"required": [
				"urn",
				"name",
				"email",
				"created",
				"updated"
			],
			"properties": {
				"urn": {
					"type": "string",
					"description": "unique urn representing a specific OC site"
				},
				"name": {
					"type": "string",
					"description": "unique identifier representing a specific OC site"
				},
				"email": {
					"type": "string",
					"format": "email",
					"description": "administrator's email address"
				},
				"related": {
					"type": "string",
					"description": "extension point for storing structured or unstructured information"
				},
				"quota": {
					"type": "integer",
					"format": "int64",
					"description": "Total number of assets this site can create"
				},
				"remQuota": {
					"type": "integer",
					"format": "int64",
					"description": "Remaining number of assets this site can create"
				},
				"created": {
					"type": "integer",
					"format": "int64",
					"description": "site registration date"
				},
				"updated": {
					"type": "integer",
					"format": "int64",
					"description": "site last update date"
				}
			}
		},
...
		"quotadesc": {
			"type": "object",
			"required": [
				"quota"
			],
			"properties": {
				"quota": {
					"type": "integer",
					"format": "int64",
					"description": "quota value"
				}
			}
		},
		"unregisteredassettypedesc": {
			"type": "object",
			"required": [
				"urn",
				"name"
			],
			"properties": {
				"urn": {
					"type": "string",
					"description": "unique urn representing a specific OC asset type"
				},
				"name": {
					"type": "string",
					"description": "unique identifier representing a specific OC asset type"
				}
			}
		},
		"unregisteredassettypeupdate": {
			"type": "object",
			"required": [
				"name"
			],
			"properties": {
				"name": {
					"type": "string",
					"description": "unique identifier representing a specific OC asset type"
				}
			}
		},
...
		"unitdesc": {
			"type": "object",
			"required": [
				"urn",
				"name",
				"description",
				"related",
				"datatype"
			],
			"properties": {
				"urn": {
					"type": "string",
					"description": "unique urn representing a specific OC unit"
				},
				"name": {
					"type": "string",
					"description": "unique identifier representing a specific OC unit"
				},
				"description": {
					"type": "string",
					"description": "Optional description"
				},
				"related": {
					"type": "string",
					"description": "extension point for storing structured or unstructured information"
				},
				"datatype": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/datatypedesc"
					}
				}
			}
		},
		"datatypedesc": {
			"type": "object",
			"required": [
				"urn",
				"name"
			],
			"properties": {
				"urn": {
					"type": "string",
					"description": "unique urn representing a specific OC data type"
				},
				"name": {
					"type": "string",
					"description": "unique identifier representing a specific OC data type"
				}
			}
		},
...
	}
}